requirements.txt
beautifulsoup4==4.4.1
future==0.15.2
google-api-python-client==1.5.0
httplib2==0.9.2
oauth2client==2.0.2
pyasn1==0.1.9
pyasn1-modules==0.0.8
python-telegram-bot==4.0.3
requests==2.10.0
rsa==3.4.2
simplejson==3.8.2
six==1.10.0
uritemplate==0.6
